+++ mmcls/models/utils/__init__.py
@@ -1,3 +1,4 @@
+from .attention import ShiftWindowMSA, WindowMSA
 from .augment.augments import Augments
 from .channel_shuffle import channel_shuffle
 from .embed import HybridEmbed, PatchEmbed, PatchMerging
@@ -9,5 +10,5 @@ from .se_layer import SELayer
 __all__ = [
     'channel_shuffle', 'make_divisible', 'InvertedResidual', 'SELayer',
     'to_ntuple', 'to_2tuple', 'to_3tuple', 'to_4tuple', 'PatchEmbed',
-    'PatchMerging', 'HybridEmbed', 'Augments'
+    'PatchMerging', 'HybridEmbed', 'Augments', 'WindowMSA', 'ShiftWindowMSA'
 ]
